+++ src/KurrentDB.SecondaryIndexing/Indexes/EventType/EventTypeIndexReader.cs
@@ -14,8 +14,7 @@ internal class EventTypeIndexReader(
 	EventTypeIndexProcessor processor,
 	IReadIndex<string> index,
 	QueryInFlightRecords<EventTypeRecord> queryInFlightRecords
-)
-	: SecondaryIndexReaderBase(index) {
+) : SecondaryIndexReaderBase(index) {
 	protected override long GetId(string streamName) {
 		if (!streamName.StartsWith(EventTypeIndex.IndexPrefix)) {
 			return ExpectedVersion.Invalid;
@@ -25,7 +24,7 @@ internal class EventTypeIndexReader(
 		return processor.GetEventTypeId(eventTypeName);
 	}
 
-	protected override long GetLastIndexedSequence(long id) => processor.GetLastEventNumber(id);
+	protected override long GetLastIndexedSequence(long id) => processor.GetLastEventNumber((int)id);
 
 	protected override IEnumerable<IndexedPrepare> GetIndexRecords(long id, long fromEventNumber, long toEventNumber) {
 		var range = db.Pool.Query<ReadEventTypeIndexQueryArgs, EventTypeRecord, ReadEventTypeIndexQuery>(new((int)id, fromEventNumber, toEventNumber));
